fixing type error

+++ myproject_host_cl.cpp
@@ -83,7 +83,7 @@ int main(int argc, char **argv) {
     }
     
     // Copying in testbench data
-    int n = std::min(inputData.size(), INSTREAMSIZE * NUM_CU * NUM_THREAD);
+    int n = std::min((int) inputData.size(), INSTREAMSIZE * NUM_CU * NUM_THREAD);
     for (int i = 0; i < n; i++) {
         fpga.source_in[i] = inputData[i];
     }
